[Feat] Add event initialization scheduler
